Updated Near proving scheme.
This scheme uses BFT finality to prove finality of random block. It implements two types of proving scheme: proofs for "epoch" blocks & proofs for randomly selected blocks.
Proofs of randomly selected blocks rely on hashes of "epoch" blocks, which are stores in the contact. For the time being, hashes are stored in .json files to simulate contracts.
The scheme takes a set of blocks, data of "epoch" blocks and hashes of "epoch" blocks to prove BFT finality of a specified block.